# mobile-sdk/openapi.yaml
openapi: 3.0.3
info:
  title: FrontFace Mobile SDK API
  version: "1.0.0"
  description: >
    Public API surface used by the FrontFace mobile chat SDK (Flutter). Covers AI chat,
    live human handoff, lead capture, and customer identify. See INTEGRATION_GUIDE.md for the
    behavioral contract (session lifecycle, realtime handoff, polling fallback, state machine).


    All endpoints are authenticated with a publishable client key in the `X-FrontFace-Key`
    header and identified by a stable per-install `X-Visitor-Id`. Live agent replies are
    delivered out-of-band over Supabase Realtime (channel `conversation:<id>`); that WebSocket
    transport is not described here — see the guide.
servers:
  - url: https://api.frontface.app
    description: Production
security:
  - FrontFaceKey: []
...
paths:
...
  /api/chat/lead-capture/status:
    get:
      tags: [Lead capture]
      summary: Has this visitor already completed the lead form?
      operationId: getLeadCaptureStatus
      parameters:
        - $ref: "#/components/parameters/VisitorIdHeader"
        - name: projectId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: visitorId
          in: query
          required: true
          schema: { type: string, maxLength: 100 }
      responses:
        "200":
          description: Completion flag
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasCompletedForm: { type: boolean }
  /api/chat/lead-capture/submit-form:
    post:
      tags: [Lead capture]
      summary: Submit the lead-capture form
      operationId: submitLeadForm
      parameters:
        - $ref: "#/components/parameters/VisitorIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadFormRequest" }
      responses:
        "200":
          description: Result (may include a qualifying question and a fresh sessionId)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadFormResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/chat/lead-capture/submit-inline:
    post:
      tags: [Lead capture]
      summary: Lightweight inline email-only capture
      operationId: submitInlineEmail
      parameters:
        - $ref: "#/components/parameters/VisitorIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, visitorId, email]
              properties:
                projectId: { type: string, format: uuid }
                visitorId: { type: string, maxLength: 100 }
                sessionId: { type: string, format: uuid, nullable: true }
                email: { type: string, format: email }
                captureSource: { type: string, example: inline_email }
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadFormResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    FrontFaceKey:
      type: apiKey
      in: header
      name: X-FrontFace-Key
      description: Publishable client key (pk_...). Safe to ship in the app binary; revocable.
  parameters:
    VisitorIdHeader:
      name: X-Visitor-Id
      in: header
      required: true
      description: Stable per-install visitor id (also mirrored into request bodies).
      schema: { type: string, maxLength: 100 }
...
  responses:
...
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Source:
      type: string
      enum: [mobile]
      default: mobile
      description: Always "mobile" for the mobile SDK.
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    LeadFormRequest:
      type: object
      required: [projectId, visitorId, formData]
      properties:
        projectId: { type: string, format: uuid }
        visitorId: { type: string, maxLength: 100 }
        sessionId: { type: string, format: uuid, nullable: true }
        source: { $ref: "#/components/schemas/Source" }
        firstMessage: { type: string, description: "The visitor's first message, or empty string." }
        formData:
          type: object
          required: [email]
          properties:
            email: { type: string, format: email }
            field_2: { $ref: "#/components/schemas/LeadFieldValue" }
            field_3: { $ref: "#/components/schemas/LeadFieldValue" }
    LeadFieldValue:
      type: object
      properties:
        label: { type: string }
        value: { type: string }
    LeadFormResponse:
      type: object
      properties:
        success: { type: boolean }
        leadId: { type: string, format: uuid }
        nextAction:
          type: string
          enum: [qualifying_question, none]
        qualifyingQuestion: { type: string }
        assembledGreeting:
          type: string
          description: Render as the next assistant bubble.
        sessionId:
          type: string
          format: uuid
          nullable: true
          description: A fresh conversation id may be returned (email_first) — persist it.
...